src/utils/commons.ts
export function dropKeys<T>(
  src: { [key: string]: T },
  keyMap: { [key: string]: any }
): { [key: string]: T } {
  return Object.keys(src).reduce<{ [key: string]: T }>((p, c) => {
    if (!(c in keyMap)) {
      p[c] = src[c]
    }
    return p
  }, {})
}

export function toKeyMap<T extends object>(
  list: T[],
  key: string | number
): { [key: string]: T } {
  return list.reduce<{ [key: string]: T }>((p, c: any) => {
    const k = c[key]
    p[k] = c
    return p
  }, {})
}

export function toList<T>(map: { [key: string]: T }): T[] {
  return Object.keys(map).map((key) => map[key])
}
